+++ ncarview/src/lib/libctrans/X11_class4.c
@@ -1,5 +1,5 @@
 /*
- *	$Id: X11_class4.c,v 1.20 1992-11-03 21:35:48 clyne Exp $
+ *	$Id: X11_class4.c,v 1.21 1993-01-06 21:11:55 clyne Exp $
  */
 /***********************************************************************
 *                                                                      *
@@ -540,7 +540,7 @@ static	int	x11_cell_array(c, color_pal, P, Q, R, nx, ny)
                 for (k=0; k<nx; k++) {
 
 			/* make sure data available in cgmc     */
-			if ((unsigned short) cgmc_index >= c->Cnum && c->more) {
+			if (cgmc_index >= c->Cnum && c->more) {
 				if (Instr_Dec(c) < 1) {
 					return (-1);
 				}
@@ -677,8 +677,9 @@ CGMC *c;
 	 * p = count of processed unsent point specifications.
 	 */
 
-	while ((unsigned short) n < c->Pnum)
+	while (n < c->Pnum)
 	{
+
 		Points.P[p].x = (short) XConvert(c->p[n].x);
 		Points.P[p].y = (short) YConvert(c->p[n].y);
 		n++;
@@ -766,7 +767,7 @@ int	X11_DisjtLine(c)
 	 * last group. n = count of processed points .
 	 * p = count of processed unsent point specifications.
 	 */
-	for(i=0,p=0; (unsigned short) i<c->Pnum; i+=2) 
+	for(i=0,p=0; i<c->Pnum; i+=2) 
 	{
 		xsegments[p].x1 = (short) XConvert(c->p[i].x);
 		xsegments[p].y1 = (short) YConvert(c->p[i].y);
@@ -823,7 +824,7 @@ int	X11_PolyMarker(c)
 
 	switch (MARKER_TYPE) {
 	case MARKER_X:	
-		for(i=0;(unsigned short) i<c->Pnum;i++) {
+		for(i=0;i<c->Pnum;i++) {
 			XDrawLine(dpy, drawable, markerGC,  
 				(int) XConvert(c->p[i].x - offset), 
 				(int) YConvert(c->p[i].y - offset ),
@@ -838,7 +839,7 @@ int	X11_PolyMarker(c)
 		}
 		break;
 	case MARKER_CIRCLE:
-		for(i=0;(unsigned short) i<c->Pnum;i++) {
+		for(i=0;i<c->Pnum;i++) {
 			XDrawArc(dpy, drawable, markerGC, 
 				(int) (XConvert(c->p[i].x) - XScale(offset)),
 				(int) (YConvert(c->p[i].y) - YScale(offset)),
@@ -848,7 +849,7 @@ int	X11_PolyMarker(c)
 		}
 		break;
 	case MARKER_STAR:
-		for(i=0;(unsigned short) i<c->Pnum;i++) {
+		for(i=0;i<c->Pnum;i++) {
 			XDrawLine(dpy, drawable, markerGC,  
 				(int) XConvert(c->p[i].x - offset), 
 				(int) YConvert(c->p[i].y),
@@ -875,7 +876,7 @@ int	X11_PolyMarker(c)
 		}
 		break;
 	case MARKER_PLUS:
-		for(i=0;(unsigned short) i<c->Pnum;i++) {
+		for(i=0;i<c->Pnum;i++) {
 			XDrawLine(dpy, drawable, markerGC,  
 				(int) XConvert(c->p[i].x - offset), 
 				(int) YConvert(c->p[i].y),
@@ -890,7 +891,7 @@ int	X11_PolyMarker(c)
 		}
 		break;
 	case MARKER_DOT:
-		for(i=0;(unsigned short) i<c->Pnum;i++) 
+		for(i=0;i<c->Pnum;i++) 
 			XDrawPoint(dpy, drawable, markerGC, 
 				(int) XConvert(c->p[i].x), 
 				(int) YConvert(c->p[i].y));
@@ -1009,7 +1010,7 @@ int	X11_Polygon(c)
 		/*
 		 *	convert VDC to X coordinates
 		 */
-		for (i=0; (unsigned short) i < c->Pnum; i++, xindex++) {
+		for (i=0; i < c->Pnum; i++, xindex++) {
 			Points.P[xindex].x = (short) XConvert(c->p[i].x);
 			Points.P[xindex].y = (short) YConvert(c->p[i].y);
 		}
