+++ ncarview/src/lib/libctrans/default.c
@@ -1,5 +1,5 @@
 /*
- *	$Id: default.c,v 1.14 1992-11-19 21:49:15 clyne Exp $
+ *	$Id: default.c,v 1.15 1993-01-06 21:12:13 clyne Exp $
  */
 /***********************************************************************
 *                                                                      *
@@ -1172,7 +1172,7 @@ CGMC *c;
 		dt->backcolr_damage = at->backcolr_access = TRUE;
 	}
 
-	for (i=0; (unsigned short)i <c->CDnum && i<=MAX_C_I; i++,color_index++){
+	for (i=0; i <c->CDnum && i<=MAX_C_I; i++,color_index++){
 		tmp = c->cd[i].red * oPtion.rgb_scale;
 		tmp = tmp > max_intensity ? max_intensity : tmp;
 		clut->ce[color_index].rgb.red = (unsigned char) tmp;
